requirements.txt
schedule==1.1.0
psycopg2==2.9.1
requests==2.25.1
google_api_python_client==2.15.0
protobuf==3.17.3
